+++ test/service/RemoteNodeService.test.ts
@@ -553,4 +553,25 @@ describe('RemoteNodeService', () => {
             },
         ]);
     });
+    const assertPeersOnInvalidUrl = async (statisticsServiceUrl: string) => {
+        presetData.statisticsServiceUrl = statisticsServiceUrl;
+        const service = new RemoteNodeService(presetData, false);
+        const peerInfos = await service.getPeerInfos();
+        // only static nodes are returned when the statistics service client fails
+        expect(peerInfos).deep.eq([
+            {
+                publicKey: 'AAAAE7EAEEAE61EF0C50B4D05931F4325F69081B1B074D31E094C4B21E8CFB3D',
+                endpoint: { host: 'someStaticPeer', port: 7900 },
+                metadata: { name: 'someStaticPeer', roles: 'Peer,Api' },
+            },
+        ]);
+    };
+
+    it('getPeerInfos unknown statisticsServiceUrl', async () => {
+        await assertPeersOnInvalidUrl('https://testnet.symbol.invalid');
+    });
+
+    it('getPeerInfos invalid statisticsServiceUrl path', async () => {
+        await assertPeersOnInvalidUrl('https://testnet.symbol.services/invalid');
+    });
 });