+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.1.0] - NEXT
+## [1.1.1] - NEXT
 
-**Milestone**: Mainnet(1.0.2.0)
+**Milestone**: Mainnet(1.0.3.0)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v1.1.1  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+- Fixed `symbol-statistics-service-typescript-fetch-client` exception handling.
+
+## [1.1.0] - Nov-05-2021
+
+**Milestone**: Mainnet(1.0.3.0)
 
 | Package          | Version | Link                                                               |
 | ---------------- | ------- | ------------------------------------------------------------------ |